Naive Status Report
Implements logging of the file it is currently trying to provide quicker
feedback to the user that it is actually running. This overwrites the
previous line. This is labeled naive because it makes no attempt to stop
threads from waiting more often for the logger lock. In practice this
can still move quite quickly though.