Add Request Templates
Adds a naive templating engine to allow specifying how db++ forms URLs
to check. Note that in recursive mode the templates are only used on the
first pass (not in successive calls).

This removes the need for the extension parameter as this can be
specified in a request template.

Currently only two variables are use for request templates:
  * `{WORD}` - this is an item from the wordlist
  * `{BASE_URL}` - this is the base_url parameter

The single default request template is `{BASE_URL}/{WORD}`, for the old
behavior of also checking directories you should add
`{BASE_URL}/{WORD}/`.