Muffle Output
The default output used to show all options and how they were
configured. This quiets the messages and shows the options that have
been configured or are almost always useful to know.

Additionally fixes the output of the `request_method` parameter since
when data was provided but no `-X` option was present it used to display
`HEAD` for the value even though it would use `POST`